+++ tests/test_response.py
@@ -118,6 +118,7 @@ def test_date4():
     with pytest.raises(MemorValidationError, match=r"Invalid value. `date` must be a datetime object that includes timezone information."):
         response = Response(message="I am fine.", date=datetime.datetime.now())
 
+
 def test_json1():
     response1 = Response(message="I am fine.", model="GPT-4", temperature=0.5, role=Role.USER, score=0.8)
     response1_json = response1.to_json()